shacl_ast::compiled
Module component

Structs

And

sh:and specifies the condition that each value node conforms to all provided shapes. This is comparable to conjunction and the logical “and” operator.

Class

The condition specified by sh:class is that each value node is a SHACL instance of a given type.

Closed

Closed Constraint Component.

Datatype

sh:datatype specifies a condition to be satisfied with regards to the datatype of each value node.

Disjoint

sh:disjoint specifies the condition that the set of value nodes is disjoint with the set of objects of the triples that have the focus node as subject and the value of sh:disjoint as predicate.

Equals

sh:equals specifies the condition that the set of all value nodes is equal to the set of objects of the triples that have the focus node as subject and the value of sh:equals as predicate.

HasValue

sh:hasValue specifies the condition that at least one value node is equal to the given RDF term.

In

sh:in specifies the condition that each value node is a member of a provided SHACL list.

LanguageIn

The condition specified by sh:languageIn is that the allowed language tags for each value node are limited by a given list of language tags.

LessThan

sh:lessThan specifies the condition that each value node is smaller than all the objects of the triples that have the focus node as subject and the value of sh:lessThan as predicate.

LessThanOrEquals

LessThanOrEquals Constraint Component.

MaxCount

sh:maxCount specifies the maximum number of value nodes that satisfy the condition.

MaxExclusive

https://www.w3.org/TR/shacl/#MaxExclusiveConstraintComponent

MaxInclusive

https://www.w3.org/TR/shacl/#MaxInclusiveConstraintComponent

MaxLength

sh:maxLength specifies the maximum string length of each value node that satisfies the condition. This can be applied to any literals and IRIs, but not to blank nodes.

MinCount

sh:minCount specifies the minimum number of value nodes that satisfy the condition. If the minimum cardinality value is 0 then this constraint is always satisfied and so may be omitted.

MinExclusive

https://www.w3.org/TR/shacl/#MinExclusiveConstraintComponent

MinInclusive

https://www.w3.org/TR/shacl/#MinInclusiveConstraintComponent

MinLength

sh:minLength specifies the minimum string length of each value node that satisfies the condition. This can be applied to any literals and IRIs, but not to blank nodes.

Node

sh:node specifies the condition that each value node conforms to the given node shape.

Nodekind

sh:nodeKind specifies a condition to be satisfied by the RDF node kind of each value node.

Not

sh:not specifies the condition that each value node cannot conform to a given shape. This is comparable to negation and the logical “not” operator.

Or

sh:or specifies the condition that each value node conforms to at least one of the provided shapes. This is comparable to disjunction and the logical “or” operator.

Pattern

sh:property can be used to specify that each value node has a given property shape.

QualifiedValueShape

QualifiedValueShape Constraint Component.

UniqueLang

The property sh:uniqueLang can be set to true to specify that no pair of value nodes may use the same language tag.

Xone

sh:or specifies the condition that each value node conforms to at least one of the provided shapes. This is comparable to disjunction and the logical “or” operator.

Enums

CompiledComponent